nuevo

+++ app/Http/Controllers/EmpleadoController.php
@@ -51,7 +51,8 @@ class EmpleadoController extends Controller
         }
 
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+        return redirect('empleado')->with('mensaje','Empleado agregado con éxito');
     }
 
     /**
@@ -113,7 +114,14 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
+
+        $empleado=Empleado::findOrFail($id);
+        
+        if(Storage::delete('public/'.$empleado->Foto)){
+
+            Empleado::destroy($id);
+
+        }    
         return redirect('empleado');
     }
 }
